fixing content bug

+++ index.php
@@ -17,18 +17,15 @@ if(isset($_COOKIE['ACTIVE_NOTE_ID'])) {
     }
 }
 
-else {
-    setcookie("ACTIVE_NOTE_ID", $db->getMaxId());
-    $activeNoteId = $db->getMaxId();
-}
-
 if(isset($_REQUEST['action'])) {
     switch($_REQUEST['action']) {
         case 'delete':
+            if($activeNoteId){
             $db->deleteNote($activeNoteId);
             $newId = $db->getMaxId();
             setcookie("ACTIVE_NOTE_ID", $newId);
             $activeNoteId = $newId;
+            }
             break;
         case 'update':
             if($activeNoteId){
